Added Nintendos Very Own, animated and all
Mario is animated with movement, needs some tweaking. Jumps have also been calculated to have an accurate height as in mario games and a letgo mechanic which lets mario jump lower than usual if you letgo early. Overall a nice change. Will add Tilesets and make the running and jumping accurate to mario games (if not the nes version)